overlooked integers from change Integer->Long

+++ bundles/org.eclipse.mita.base/src/org/eclipse/mita/base/expressions/terminals/ExpressionsValueConverterService.java
@@ -58,12 +58,12 @@ public class ExpressionsValueConverterService extends DefaultTerminalConverters 
 	}
 	
 	@ValueConverter(rule = HEX)
-	public IValueConverter<Integer> HEX() {
+	public IValueConverter<Long> HEX() {
 		return hexConverter;
 	}
 	
 	@ValueConverter(rule = BINARY)
-	public IValueConverter<Integer> BINARY() {
+	public IValueConverter<Long> BINARY() {
 		return binaryConverter;
 	}
